pyproject.toml
[project]
name = "acs-nativity-streamlit"
version = "0.1.0"
description = "Streamlit app that lets people interact with nativity data generated by the `acs-nativity` package."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "acs-nativity>=0.1.1",
    "streamlit>=1.55.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "mypy>=1.19.1",
    "pandas-stubs>=3.0.0.260204",
    "ruff>=0.15.6",
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
